refactor : 할인 계산 전략패턴 활용
