+++ src/main/java/christmas/model/strategy/DiscountStrategy.java
@@ -1,0 +1,10 @@
+package christmas.model.strategy;
+
+import christmas.model.Date;
+import christmas.model.MenuCount;
+
+public interface DiscountStrategy {
+    boolean canGetDiscount(MenuCount menuCount);
+    int calculatePromotionAmount(MenuCount menuCount, Date date);
+    int calaulateTotalDiscountAmount(MenuCount menuCount, Date date);
+}
