+++ src/test/java/christmas/DiscountTest.java
@@ -1,9 +1,8 @@
 package christmas;
 
 import christmas.model.Date;
-import christmas.model.Menu;
 import christmas.model.MenuCount;
-import christmas.service.DiscountService;
+import christmas.model.strategy.WootecoDiscountStrategy;
 import christmas.util.Parser;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -13,11 +12,11 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 public class DiscountTest {
-    private DiscountService discountService;
+    private WootecoDiscountStrategy discountStrategy;
 
     @BeforeEach
     void setUp() {
-        discountService = new DiscountService();
+        discountStrategy = new WootecoDiscountStrategy();
     }
 
     @DisplayName("크리스마스 날짜 전이면 크리스마스 할인을 받는다.")
@@ -26,7 +25,7 @@ public class DiscountTest {
         //given
         Date date = Date.of(23);
         // when & then
-        Assertions.assertThat(discountService.christmasDiscount(date)).isEqualTo(3200);
+        Assertions.assertThat(discountStrategy.christmasDiscount(date)).isEqualTo(3200);
     }
 
     @DisplayName("크리스마스 날짜 후 이면 크리스마스 할인을 못 받는다.")
@@ -35,7 +34,7 @@ public class DiscountTest {
         //given
         Date date = Date.of(26);
         // when & then
-        Assertions.assertThat(discountService.christmasDiscount(date)).isEqualTo(0);
+        Assertions.assertThat(discountStrategy.christmasDiscount(date)).isEqualTo(0);
     }
 
     @DisplayName("주말이면 메인 메뉴 1개당 2,023원 할인을 받는다.")
@@ -46,7 +45,7 @@ public class DiscountTest {
         String input = "티본스테이크-1,바비큐립-1,초코케이크-2,제로콜라-1";
         MenuCount menuCount = new MenuCount(Parser.parseMenuCount(input));
         // when & then
-        Assertions.assertThat(discountService.weekendDiscount(menuCount, date)).isEqualTo(4046);
+        Assertions.assertThat(discountStrategy.weekendDiscount(menuCount, date)).isEqualTo(4046);
     }
 
     @DisplayName("평일이면 디저트 메뉴 1개당 2,023원 할인을 받는다.")
@@ -57,7 +56,7 @@ public class DiscountTest {
         String input = "티본스테이크-1,바비큐립-1,초코케이크-2,제로콜라-1";
         MenuCount menuCount = new MenuCount(Parser.parseMenuCount(input));
         // when & then
-        Assertions.assertThat(discountService.weekdayDiscount(menuCount, date)).isEqualTo(4046);
+        Assertions.assertThat(discountStrategy.weekdayDiscount(menuCount, date)).isEqualTo(4046);
     }
 
     @DisplayName("특별한 날짜이면 총주문 금액에서 1000원 할인을 받는다.")
@@ -66,7 +65,7 @@ public class DiscountTest {
         //given
         Date date = Date.of(3);
         // when & then
-        Assertions.assertThat(discountService.specialDayDiscount(date)).isEqualTo(1000);
+        Assertions.assertThat(discountStrategy.specialDayDiscount(date)).isEqualTo(1000);
     }
 
     @DisplayName("총주문 금액이 120,000원 이상이면 증정품을 받는다.")
@@ -76,7 +75,7 @@ public class DiscountTest {
         String input = "티본스테이크-1,바비큐립-1,초코케이크-2,제로콜라-1";
         MenuCount menuCount = new MenuCount(Parser.parseMenuCount(input));
         // when & then
-        Assertions.assertThat(discountService.giftEventDiscount(menuCount)).isEqualTo(25000);
+        Assertions.assertThat(discountStrategy.giftEventDiscount(menuCount)).isEqualTo(25000);
     }
 
     @DisplayName("총할인 금액을 계산한다.")
@@ -90,7 +89,7 @@ public class DiscountTest {
         Date date = Date.of(inputDate);
         MenuCount menuCount = new MenuCount(Parser.parseMenuCount(input));
         // when & then
-        Assertions.assertThat(discountService.calculateTotalDiscountAmount(menuCount, date))
+        Assertions.assertThat(discountStrategy.calaulateTotalDiscountAmount(menuCount, date))
                 .isEqualTo(expectedAmount);
     }
 
@@ -105,7 +104,7 @@ public class DiscountTest {
         Date date = Date.of(inputDate);
         MenuCount menuCount = new MenuCount(Parser.parseMenuCount(input));
         // when & then
-        Assertions.assertThat(discountService.calculatePromotionAmount(menuCount, date))
+        Assertions.assertThat(discountStrategy.calculatePromotionAmount(menuCount, date))
                 .isEqualTo(expectedAmount);
     }
 
@@ -120,7 +119,7 @@ public class DiscountTest {
         Date date = Date.of(inputDate);
         MenuCount menuCount = new MenuCount(Parser.parseMenuCount(input));
         // when & then
-        Assertions.assertThat(discountService.calculatePromotionAmount(menuCount, date))
+        Assertions.assertThat(discountStrategy.calculatePromotionAmount(menuCount, date))
                 .isEqualTo(expectedAmount);
     }
